image/spec: add missing "ArgsEscaped" field (v1.1, v1.2)
This field was added in 9db5db1b94bc1000d151315851096dcc6cd3512d, which
was part of v1.10.0-rc1 and later, which used image spec v1.1.0.

It's worth noting that documentation for the v1.1.0 image spec was not
yet available until commit 1db06cf8ecb870988f57da59a10c29a3fe5f83c5,
which was included in v1.12.0-rc1 and up. The `ArgsEscaped` field was
also adopted by the OCI image spec since [v1.1.0-rc3][1], but considered
deprecated, and not recommended to be used.

This patch amends the v1.1 and v1.2 specifications to describe the field.

[1]: https://github.com/opencontainers/image-spec/commit/59780aa56914ae9cf4ca68c10989d3f89337d1e0

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>