Rename package specs & retract v1.3.0 tag
The package at package path github.com/moby/docker-image-spec/specs-go
was assigned the package name 'v1'. It is supposed to have the package
name 'specs' to mirror the name of the
github.com/opencontainers/image-spec/specs-go package, but this mistake
was not caught until after stable module versions were tagged. Fix the
package name and retract the v1.3.0 tag.

Signed-off-by: Cory Snider <csnider@mirantis.com>